Update installs to be correct. Better testing for existance, moving things to the production vs non-production install

+++ chromatik/installations/install.ps1
@@ -28,9 +28,12 @@ python shrub.py --config $install_dir/shrubs.json --fixtures_folder $fixtures_di
 python tree.py --tree_config $install_dir/trees.json --branch_config $install_dir/tree_branches.csv --fixtures_folder $fixtures_dir
 python bench.py --config $install_dir/bench.json --fixtures_folder $fixtures_dir
 cp $install_dir/entwined.lxp "$HOME/Chromatik/Projects"
-cp $install_dir/config.json "$HOME/Chromatik"
 
-if (Get-Item -Path $install_dir/autoplay.lxr -ErrorAction Ignore) {
+if (Test-Path -Path $install_dir/config.json -PathType Leaf) {
+        cp $install_dir/config.json "$HOME/Chromatik"
+}
+
+if (Test-Path -Path $install_dir/autoplay.lxr -PathType Leaf) {
         cp $install_dir/autoplay.lxr "$HOME/Chromatik" 
 }
 else {
